Introduced generic data pipeline (to be used by all models) and model specific pipelines Created sklearn preprocessing and training pipelines for logistic regression Wrapped data cleaner into a sklearn transformer class to be used as one in a pipeline Implemented partially train_model.py as the orchestrator to train a model Major bug fixes

+++ src/features/feature_engineering.py
@@ -18,7 +18,7 @@ def build_derived_features(df: pd.DataFrame) -> pd.DataFrame:
     df[nii_ratio] = df[nii] / total
     df[nii_ratio] = df[nii_ratio].fillna(0)
 
-    df[is_gmp_missing] = df[gmp].isna()
+    df[is_gmp_missing] = df[gmp].isna().astype(int)
     df[gmp] = df[gmp].fillna(0)
 
     return df
